client-api: Move `RoomPowerLevelsContentOverride` to `create_room::v3`
Make it `JsonCastable` to `RoomPowerLevelsEventContent` so we can use it without a breaking change.
