+++ crates/ruma-client-api/CHANGELOG.md
@@ -1,12 +1,12 @@
 # [unreleased]
 
-Breaking changes:
+Improvements:
 
-- `create_room::v3::Request::power_level_content_override` is now of type 
-  `Option<Raw<RoomPowerLevelsContentOverride>>`. This `RoomPowerLevelsContentOverride` 
-  type has the same signature as `RoomPowerLevelsEventContent` except all `Int` 
-  fields are `Option<Int>`, which allows uploading explicit values no matter what the 
-  default ones are.
+- Added `create_room::v3::RoomPowerLevelsContentOverride`. This has the same 
+  signature as `RoomPowerLevelsEventContent` except all `Int` fields are 
+  `Option<Int>`, which allows uploading explicit values no matter what the 
+  default ones are. When in `Raw` form, this type can be casted to
+  `RoomPowerLevelsEventContent` and used to populate `power_level_content_override`.
 
 # 0.22.1
