+++ crates/ruma-client-api/src/room/create_room.rs
@@ -2,6 +2,13 @@
 //!
 //! Create a new room.
 
+use std::collections::BTreeMap;
+
+use js_int::Int;
+use ruma_common::{OwnedUserId, power_levels::NotificationPowerLevels, serde::JsonCastable};
+use ruma_events::{TimelineEventType, room::power_levels::RoomPowerLevelsEventContent};
+use serde::Serialize;
+
 pub mod v3 {
     //! `/v3/` ([spec])
     //!
@@ -19,7 +26,7 @@ pub mod v3 {
         AnyInitialStateEvent,
         room::{
             create::{PreviousRoom, RoomCreateEventContent},
-            power_levels::RoomPowerLevelsContentOverride,
+            power_levels::RoomPowerLevelsEventContent,
         },
     };
     use serde::{Deserialize, Serialize};
@@ -71,7 +78,7 @@ pub mod v3 {
 
         /// Power level content to override in the default power level event.
         #[serde(skip_serializing_if = "Option::is_none")]
-        pub power_level_content_override: Option<Raw<RoomPowerLevelsContentOverride>>,
+        pub power_level_content_override: Option<Raw<RoomPowerLevelsEventContent>>,
 
         /// Convenience parameter for setting various default state events based on a preset.
         #[serde(skip_serializing_if = "Option::is_none")]
@@ -210,5 +217,133 @@ pub mod v3 {
 
         #[doc(hidden)]
         _Custom(PrivOwnedStr),
+    }
+}
+
+/// The power level values that can be overridden when creating a room.
+#[derive(Clone, Debug, Serialize, Default)]
+#[cfg_attr(not(ruma_unstable_exhaustive_types), non_exhaustive)]
+pub struct RoomPowerLevelsContentOverride {
+    /// The level required to ban a user.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub ban: Option<Int>,
+
+    /// The level required to send specific event types.
+    ///
+    /// This is a mapping from event type to power level required.
+    #[serde(default, skip_serializing_if = "BTreeMap::is_empty")]
+    pub events: BTreeMap<TimelineEventType, Int>,
+
+    /// The default level required to send message events.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub events_default: Option<Int>,
+
+    /// The level required to invite a user.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub invite: Option<Int>,
+
+    /// The level required to kick a user.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub kick: Option<Int>,
+
+    /// The level required to redact an event.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub redact: Option<Int>,
+
+    /// The default level required to send state events.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub state_default: Option<Int>,
+
+    /// The power levels for specific users.
+    ///
+    /// This is a mapping from `user_id` to power level for that user.
+    #[serde(default, skip_serializing_if = "BTreeMap::is_empty")]
+    pub users: BTreeMap<OwnedUserId, Int>,
+
+    /// The default power level for every user in the room.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub users_default: Option<Int>,
+
+    /// The power level requirements for specific notification types.
+    ///
+    /// This is a mapping from `key` to power level for that notifications key.
+    #[serde(default, skip_serializing_if = "NotificationPowerLevels::is_default")]
+    pub notifications: NotificationPowerLevels,
+}
+
+impl JsonCastable<RoomPowerLevelsEventContent> for RoomPowerLevelsContentOverride {}
+
+#[cfg(test)]
+mod tests {
+    use std::collections::BTreeMap;
+
+    use assign::assign;
+    use js_int::int;
+    use maplit::btreemap;
+    use ruma_common::{power_levels::NotificationPowerLevels, user_id};
+    use serde_json::json;
+
+    #[test]
+    fn serialization_of_power_levels_overridden_values_with_optional_fields_as_none() {
+        let power_levels = RoomPowerLevelsContentOverride {
+            ban: None,
+            events: BTreeMap::new(),
+            events_default: None,
+            invite: None,
+            kick: None,
+            redact: None,
+            state_default: None,
+            users: BTreeMap::new(),
+            users_default: None,
+            notifications: NotificationPowerLevels::default(),
+        };
+
+        let actual = to_json_value(&power_levels).unwrap();
+        let expected = json!({});
+
+        assert_eq!(actual, expected);
+    }
+
+    #[test]
+    fn serialization_of_power_levels_overridden_values_with_all_fields() {
+        let user = user_id!("@carl:example.com");
+        let power_levels_event = RoomPowerLevelsContentOverride {
+            ban: Some(int!(23)),
+            events: btreemap! {
+                "m.dummy".into() => int!(23)
+            },
+            events_default: Some(int!(23)),
+            invite: Some(int!(23)),
+            kick: Some(int!(23)),
+            redact: Some(int!(23)),
+            state_default: Some(int!(23)),
+            users: btreemap! {
+                user.to_owned() => int!(23)
+            },
+            users_default: Some(int!(23)),
+            notifications: assign!(NotificationPowerLevels::new(), { room: int!(23) }),
+        };
+
+        let actual = to_json_value(&power_levels_event).unwrap();
+        let expected = json!({
+            "ban": 23,
+            "events": {
+                "m.dummy": 23
+            },
+            "events_default": 23,
+            "invite": 23,
+            "kick": 23,
+            "redact": 23,
+            "state_default": 23,
+            "users": {
+                "@carl:example.com": 23
+            },
+            "users_default": 23,
+            "notifications": {
+                "room": 23
+            },
+        });
+
+        assert_eq!(actual, expected);
     }
 }
