+++ crates/ruma-events/src/room/power_levels.rs
@@ -923,89 +923,6 @@ pub enum PowerLevelUserAction {
     ChangePowerLevel,
 }
 
-/// The power level values that can be overridden when creating a room.
-#[derive(Clone, Debug, Serialize)]
-#[cfg_attr(not(ruma_unstable_exhaustive_types), non_exhaustive)]
-pub struct RoomPowerLevelsContentOverride {
-    /// The level required to ban a user.
-    #[serde(skip_serializing_if = "ruma_common::serde::is_default")]
-    pub ban: Option<Int>,
-
-    /// The level required to send specific event types.
-    ///
-    /// This is a mapping from event type to power level required.
-    #[serde(
-        default,
-        skip_serializing_if = "BTreeMap::is_empty",
-        deserialize_with = "ruma_common::serde::btreemap_deserialize_v1_powerlevel_values"
-    )]
-    pub events: BTreeMap<TimelineEventType, Int>,
-
-    /// The default level required to send message events.
-    #[serde(skip_serializing_if = "ruma_common::serde::is_default")]
-    pub events_default: Option<Int>,
-
-    /// The level required to invite a user.
-    #[serde(skip_serializing_if = "ruma_common::serde::is_default")]
-    pub invite: Option<Int>,
-
-    /// The level required to kick a user.
-    #[serde(skip_serializing_if = "ruma_common::serde::is_default")]
-    pub kick: Option<Int>,
-
-    /// The level required to redact an event.
-    #[serde(skip_serializing_if = "ruma_common::serde::is_default")]
-    pub redact: Option<Int>,
-
-    /// The default level required to send state events.
-    #[serde(skip_serializing_if = "ruma_common::serde::is_default")]
-    pub state_default: Option<Int>,
-
-    /// The power levels for specific users.
-    ///
-    /// This is a mapping from `user_id` to power level for that user.
-    #[serde(
-        default,
-        skip_serializing_if = "BTreeMap::is_empty",
-        deserialize_with = "ruma_common::serde::btreemap_deserialize_v1_powerlevel_values"
-    )]
-    pub users: BTreeMap<OwnedUserId, Int>,
-
-    /// The default power level for every user in the room.
-    #[serde(skip_serializing_if = "ruma_common::serde::is_default")]
-    pub users_default: Option<Int>,
-
-    /// The power level requirements for specific notification types.
-    ///
-    /// This is a mapping from `key` to power level for that notifications key.
-    #[serde(default, skip_serializing_if = "NotificationPowerLevels::is_default")]
-    pub notifications: NotificationPowerLevels,
-}
-
-impl RoomPowerLevelsContentOverride {
-    /// Creates a new, empty instance of [`RoomPowerLevelsContentOverride`].
-    pub fn new() -> Self {
-        Self {
-            ban: None,
-            events: Default::default(),
-            events_default: None,
-            invite: None,
-            kick: None,
-            redact: None,
-            state_default: None,
-            users: Default::default(),
-            users_default: None,
-            notifications: Default::default(),
-        }
-    }
-}
-
-impl Default for RoomPowerLevelsContentOverride {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use std::collections::BTreeMap;
@@ -1017,8 +934,8 @@ mod tests {
     use serde_json::{json, to_value as to_json_value};
 
     use super::{
-        NotificationPowerLevels, RoomPowerLevels, RoomPowerLevelsContentOverride,
-        RoomPowerLevelsEventContent, RoomPowerLevelsSource, default_power_level,
+        NotificationPowerLevels, RoomPowerLevels, RoomPowerLevelsEventContent,
+        RoomPowerLevelsSource, default_power_level,
     };
 
     #[test]
@@ -1061,70 +978,6 @@ mod tests {
                 user.to_owned() => int!(23)
             },
             users_default: int!(23),
-            notifications: assign!(NotificationPowerLevels::new(), { room: int!(23) }),
-        };
-
-        let actual = to_json_value(&power_levels_event).unwrap();
-        let expected = json!({
-            "ban": 23,
-            "events": {
-                "m.dummy": 23
-            },
-            "events_default": 23,
-            "invite": 23,
-            "kick": 23,
-            "redact": 23,
-            "state_default": 23,
-            "users": {
-                "@carl:example.com": 23
-            },
-            "users_default": 23,
-            "notifications": {
-                "room": 23
-            },
-        });
-
-        assert_eq!(actual, expected);
-    }
-
-    #[test]
-    fn serialization_of_overridden_values_with_optional_fields_as_none() {
-        let power_levels = RoomPowerLevelsContentOverride {
-            ban: None,
-            events: BTreeMap::new(),
-            events_default: None,
-            invite: None,
-            kick: None,
-            redact: None,
-            state_default: None,
-            users: BTreeMap::new(),
-            users_default: None,
-            notifications: NotificationPowerLevels::default(),
-        };
-
-        let actual = to_json_value(&power_levels).unwrap();
-        let expected = json!({});
-
-        assert_eq!(actual, expected);
-    }
-
-    #[test]
-    fn serialization_of_overridden_values_with_all_fields() {
-        let user = user_id!("@carl:example.com");
-        let power_levels_event = RoomPowerLevelsContentOverride {
-            ban: Some(int!(23)),
-            events: btreemap! {
-                "m.dummy".into() => int!(23)
-            },
-            events_default: Some(int!(23)),
-            invite: Some(int!(23)),
-            kick: Some(int!(23)),
-            redact: Some(int!(23)),
-            state_default: Some(int!(23)),
-            users: btreemap! {
-                user.to_owned() => int!(23)
-            },
-            users_default: Some(int!(23)),
             notifications: assign!(NotificationPowerLevels::new(), { room: int!(23) }),
         };
